Replace alert keys for mainnet with activation keys.
Note: seperate alert keys have been removed and replaced with the keys used to sign activations.

Note: Michael is no longer a regular contributor to the project, and as such his key has been removed.

Note: Adam is a project maintainer and is a signatory to activations, and as such his key as been added.