Fix fail safe iteration when forming ECDSA point
The last byte of a pubkey of Class B transactions is iterated until a valid ECDSA point is formed. A valid point is usually right around the corner, but there is a fail safe included to avoid looping the last byte forever, in case there is none. The loop for this was faulty and is fixed now.

Thanks to @lxjxiaojun for spotting this!